requirements.txt
numpy>=1.21.6
onnx>=1.13.1
opencv_python>=4.5.1.48
Pillow>=8.1.2
scikit_learn>=0.24.2
tqdm>=4.65.0
tensorboard
pyyaml